# dummy-api/gen/http/dummy/server/encode_decode.go
// Code generated by goa v3.23.4, DO NOT EDIT.
//
// dummy HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/vidwadeseram/go-boilerplate/dummy-api/design

package server

import (
	"context"
	"errors"
	"io"
	"net/http"

	dummy "github.com/vidwadeseram/go-boilerplate/dummy-api/gen/dummy"
	dummyviews "github.com/vidwadeseram/go-boilerplate/dummy-api/gen/dummy/views"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeCreateItemResponse returns an encoder for responses returned by the
// dummy create_item endpoint.
func EncodeCreateItemResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res := v.(*dummyviews.Item)
		enc := encoder(ctx, w)
		body := NewCreateItemResponseBody(res.Projected)
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeCreateItemRequest returns a decoder for requests sent to the dummy
// create_item endpoint.
func DecodeCreateItemRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*dummy.CreateItemPayload, error) {
	return func(r *http.Request) (*dummy.CreateItemPayload, error) {
		var (
			body CreateItemRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if errors.Is(err, io.EOF) {
				return nil, goa.MissingPayloadError()
			}
			var gerr *goa.ServiceError
			if errors.As(err, &gerr) {
				return nil, gerr
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateCreateItemRequestBody(&body)
		if err != nil {
			return nil, err
		}

		var (
			token string
		)
		token = r.Header.Get("Authorization")
		if token == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("token", "header"))
		}
		if err != nil {
			return nil, err
		}
		payload := NewCreateItemPayload(&body, token)

		return payload, nil
	}
}

// EncodeListItemsResponse returns an encoder for responses returned by the
// dummy list_items endpoint.
func EncodeListItemsResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*dummy.ItemsCollection)
		enc := encoder(ctx, w)
		body := NewListItemsResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeListItemsRequest returns a decoder for requests sent to the dummy
// list_items endpoint.
func DecodeListItemsRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*dummy.ListItemsPayload, error) {
	return func(r *http.Request) (*dummy.ListItemsPayload, error) {
		var (
			token string
			err   error
		)
		token = r.Header.Get("Authorization")
		if token == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("token", "header"))
		}
		if err != nil {
			return nil, err
		}
		payload := NewListItemsPayload(token)

		return payload, nil
	}
}

// EncodeGetItemResponse returns an encoder for responses returned by the dummy
// get_item endpoint.
func EncodeGetItemResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res := v.(*dummyviews.Item)
		enc := encoder(ctx, w)
		body := NewGetItemResponseBody(res.Projected)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetItemRequest returns a decoder for requests sent to the dummy
// get_item endpoint.
func DecodeGetItemRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*dummy.ItemIDPayload, error) {
	return func(r *http.Request) (*dummy.ItemIDPayload, error) {
		var (
			id    string
			token string
			err   error

			params = mux.Vars(r)
		)
		id = params["id"]
		token = r.Header.Get("Authorization")
		if token == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("token", "header"))
		}
		if err != nil {
			return nil, err
		}
		payload := NewGetItemItemIDPayload(id, token)

		return payload, nil
	}
}

// EncodeDeleteItemResponse returns an encoder for responses returned by the
// dummy delete_item endpoint.
func EncodeDeleteItemResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		w.WriteHeader(http.StatusNoContent)
		return nil
	}
}

// DecodeDeleteItemRequest returns a decoder for requests sent to the dummy
// delete_item endpoint.
func DecodeDeleteItemRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*dummy.ItemIDPayload, error) {
	return func(r *http.Request) (*dummy.ItemIDPayload, error) {
		var (
			id    string
			token string
			err   error

			params = mux.Vars(r)
		)
		id = params["id"]
		token = r.Header.Get("Authorization")
		if token == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("token", "header"))
		}
		if err != nil {
			return nil, err
		}
		payload := NewDeleteItemItemIDPayload(id, token)

		return payload, nil
	}
}

// marshalDummyItemToItemResponseBody builds a value of type *ItemResponseBody
// from a value of type *dummy.Item.
func marshalDummyItemToItemResponseBody(v *dummy.Item) *ItemResponseBody {
	res := &ItemResponseBody{
		ID:          v.ID,
		Name:        v.Name,
		Description: v.Description,
		OwnerID:     v.OwnerID,
		CreatedAt:   v.CreatedAt,
	}

	return res
}