# dummy-api/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Dummy Service
    description: Reference CRUD microservice that enforces identity auth
    version: 0.0.1
servers:
    - url: http://localhost:8082
paths:
    /openapi.json:
        get:
            tags:
                - dummy
            summary: Download gen/http/openapi.json
            operationId: dummy#/openapi.json
            responses:
                "200":
                    description: File downloaded
    /v1/dummy/items:
        get:
            tags:
                - dummy
            summary: list_items dummy
            operationId: dummy#list_items
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ItemsCollection'
                            example:
                                items:
                                    - created_at: "2002-12-08T21:09:38Z"
                                      description: Odio voluptas veritatis in tempore consequatur.
                                      id: Voluptatem est et eius dignissimos asperiores doloribus.
                                      name: Velit laudantium temporibus magni est.
                                      owner_id: Aliquam id aut itaque et.
                                    - created_at: "2002-12-08T21:09:38Z"
                                      description: Odio voluptas veritatis in tempore consequatur.
                                      id: Voluptatem est et eius dignissimos asperiores doloribus.
                                      name: Velit laudantium temporibus magni est.
                                      owner_id: Aliquam id aut itaque et.
                                    - created_at: "2002-12-08T21:09:38Z"
                                      description: Odio voluptas veritatis in tempore consequatur.
                                      id: Voluptatem est et eius dignissimos asperiores doloribus.
                                      name: Velit laudantium temporibus magni est.
                                      owner_id: Aliquam id aut itaque et.
                                    - created_at: "2002-12-08T21:09:38Z"
                                      description: Odio voluptas veritatis in tempore consequatur.
                                      id: Voluptatem est et eius dignissimos asperiores doloribus.
                                      name: Velit laudantium temporibus magni est.
                                      owner_id: Aliquam id aut itaque et.
        post:
            tags:
                - dummy
            summary: create_item dummy
            operationId: dummy#create_item
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateItemPayload2'
                        example:
                            description: Sunt iusto.
                            name: Quisquam odio sit eaque aut perferendis temporibus.
            responses:
                "201":
                    description: Created response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DummyItem'
                            example:
                                created_at: "1996-01-04T02:25:33Z"
                                description: Dolore adipisci est.
                                id: Saepe officiis eos.
                                name: Tempora rerum odit nam odio qui.
                                owner_id: Deleniti delectus sint.
    /v1/dummy/items/{id}:
        delete:
            tags:
                - dummy
            summary: delete_item dummy
            operationId: dummy#delete_item
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                    example: Quo in temporibus esse.
                  example: Quae ut illo occaecati doloribus et et.
            responses:
                "204":
                    description: No Content response.
        get:
            tags:
                - dummy
            summary: get_item dummy
            operationId: dummy#get_item
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                    example: Itaque aut et totam non atque.
                  example: Autem voluptates aut ea.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DummyItem'
                            example:
                                created_at: "1991-02-10T13:42:40Z"
                                description: Tenetur voluptatem sit.
                                id: Neque eos dolor id corrupti nostrum.
                                name: Aut animi qui.
                                owner_id: Ipsum dignissimos.
components:
    schemas:
        AuthenticatedPayload:
            type: object
            properties:
                token:
                    type: string
                    description: Bearer token
                    example: Sit ut.
            example:
                token: Modi quia et.
            required:
                - token
        CreateItemPayload:
            type: object
            properties:
                description:
                    type: string
                    example: Recusandae omnis.
                name:
                    type: string
                    example: Provident a adipisci possimus.
                token:
                    type: string
                    description: Bearer token
                    example: Fuga aut et qui blanditiis excepturi assumenda.
            example:
                description: Quod itaque magni eveniet impedit.
                name: Et velit repellat et accusamus qui.
                token: Dicta repellendus vel delectus quia qui.
            required:
                - name
                - token
        CreateItemPayload2:
            type: object
            properties:
                description:
                    type: string
                    example: Quidem et quibusdam illum.
                name:
                    type: string
                    example: Distinctio nemo odio cupiditate expedita dolor nostrum.
            example:
                description: Odio deserunt eos.
                name: Corrupti sapiente qui voluptatem.
            required:
                - name
        DummyItem:
            type: object
            properties:
                created_at:
                    type: string
                    example: "1998-08-27T03:10:26Z"
                    format: date-time
                description:
                    type: string
                    example: Tenetur debitis sit pariatur consequatur soluta ullam.
                id:
                    type: string
                    description: Item identifier
                    example: Voluptatum quaerat.
                name:
                    type: string
                    example: Distinctio qui.
                owner_id:
                    type: string
                    example: Dolore tempore velit alias voluptas ut.
            example:
                created_at: "1976-04-09T21:17:50Z"
                description: Id est.
                id: Officiis temporibus.
                name: Laudantium quas vero.
                owner_id: Rerum assumenda ex dolorem.
            required:
                - id
                - name
                - owner_id
                - created_at
        DummyNotFoundError:
            type: object
            properties:
                message:
                    type: string
                    example: Quia facere laboriosam ut maxime consequuntur.
            example:
                message: Aut quas et aliquid repudiandae id.
            required:
                - message
        DummyUnauthorizedError:
            type: object
            properties:
                message:
                    type: string
                    example: Commodi sit quam officia dolor natus.
            example:
                message: Facilis exercitationem quam aut enim laboriosam non.
            required:
                - message
        ItemIDPayload:
            type: object
            properties:
                id:
                    type: string
                    example: Necessitatibus esse eum quidem rerum consequatur perferendis.
                token:
                    type: string
                    description: Bearer token
                    example: Non quaerat.
            example:
                id: Placeat cum vitae ab.
                token: Et et ullam dolores nobis nostrum similique.
            required:
                - id
                - token
        ItemsCollection:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/DummyItem'
                    example:
                        - created_at: "2002-12-08T21:09:38Z"
                          description: Odio voluptas veritatis in tempore consequatur.
                          id: Voluptatem est et eius dignissimos asperiores doloribus.
                          name: Velit laudantium temporibus magni est.
                          owner_id: Aliquam id aut itaque et.
                        - created_at: "2002-12-08T21:09:38Z"
                          description: Odio voluptas veritatis in tempore consequatur.
                          id: Voluptatem est et eius dignissimos asperiores doloribus.
                          name: Velit laudantium temporibus magni est.
                          owner_id: Aliquam id aut itaque et.
                        - created_at: "2002-12-08T21:09:38Z"
                          description: Odio voluptas veritatis in tempore consequatur.
                          id: Voluptatem est et eius dignissimos asperiores doloribus.
                          name: Velit laudantium temporibus magni est.
                          owner_id: Aliquam id aut itaque et.
            example:
                items:
                    - created_at: "2002-12-08T21:09:38Z"
                      description: Odio voluptas veritatis in tempore consequatur.
                      id: Voluptatem est et eius dignissimos asperiores doloribus.
                      name: Velit laudantium temporibus magni est.
                      owner_id: Aliquam id aut itaque et.
                    - created_at: "2002-12-08T21:09:38Z"
                      description: Odio voluptas veritatis in tempore consequatur.
                      id: Voluptatem est et eius dignissimos asperiores doloribus.
                      name: Velit laudantium temporibus magni est.
                      owner_id: Aliquam id aut itaque et.
                    - created_at: "2002-12-08T21:09:38Z"
                      description: Odio voluptas veritatis in tempore consequatur.
                      id: Voluptatem est et eius dignissimos asperiores doloribus.
                      name: Velit laudantium temporibus magni est.
                      owner_id: Aliquam id aut itaque et.
            required:
                - items
        ListItemsPayload:
            type: object
            properties:
                token:
                    type: string
                    description: Bearer token
                    example: Sit et.
            example:
                token: Cupiditate odit aut optio quod.
            required:
                - token
tags:
    - name: dummy
      description: CRUD operations on items that rely on identity-api for auth